Fixed to initialize memory otherwise assertion errors occur.\nConfigured build using Visual C++ 2010.
git-svn-id: https://xemmai.googlecode.com/svn/trunk@125 21b6bbce-954d-0410-9621-597f947845d9

+++ src/object.cc
@@ -373,10 +373,10 @@ void t_object::f_field_put(t_object* a_key, const t_transfer& a_value)
 	int index = v_structure->f_index(a_key);
 	if (index < 0) {
 		index = v_structure->f_size();
-		if (!v_fields || index >= v_fields->f_size()) {
+		if (!v_fields || static_cast<size_t>(index) >= v_fields->f_size()) {
 			t_transfer tuple = t_tuple::f_instantiate(index + 1);
 			t_tuple& fields = f_as<t_tuple&>(tuple);
-			for (size_t i = 0; i < index; ++i) fields[i] = (*v_fields)[i].f_transfer();
+			for (size_t i = 0; i < static_cast<size_t>(index); ++i) fields[i] = (*v_fields)[i].f_transfer();
 			tuple.f_pointer__(v_fields);
 			v_fields = &fields;
 		}
